refactor: add Events to Core
